requirements.txt
pymongo
psycopg2-binary
pandas
python-dotenv
